requirements.txt
requests>=2.32,<3
beautifulsoup4>=4.12,<5
openai>=2.20,<3
python-dotenv>=1.0,<2
playwright>=1.49,<2
